requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
distlib==0.3.3
filelock==3.3.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
mysqlclient==2.0.3
platformdirs==2.4.0
protobuf==3.18.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.2.0
python-dotenv==0.19.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
virtualenv==20.8.1
Werkzeug==2.0.2
